+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConnectivityApi24Test.kt
@@ -1,5 +1,3 @@
-@file:Suppress("DEPRECATION")
-
 package com.bugsnag.android
 
 import android.content.Context
@@ -31,7 +29,7 @@ class ConnectivityApi24Test {
     lateinit var capabilities: NetworkCapabilities
 
     @Test
-    fun connectivityLegacyHasConnection() {
+    fun connectivityApi24HasConnection() {
         val conn = ConnectivityApi24(cm, null)
         assertFalse(conn.hasNetworkConnection())
         Mockito.`when`(cm.activeNetwork).thenReturn(info)
@@ -39,7 +37,7 @@ class ConnectivityApi24Test {
     }
 
     @Test
-    fun connectivityLegacyNetworkState() {
+    fun connectivityApi24NetworkState() {
         val conn = ConnectivityApi24(cm, null)
         Mockito.`when`(cm.activeNetwork).thenReturn(info)
 
@@ -49,5 +47,41 @@ class ConnectivityApi24Test {
         Mockito.`when`(cm.getNetworkCapabilities(info)).thenReturn(capabilities)
         Mockito.`when`(capabilities.hasTransport(0)).thenReturn(true)
         assertEquals("cellular", conn.retrieveNetworkAccessState())
+    }
+
+    @Test
+    fun connectivityApi24OnAvailable() {
+        var count = 0
+        val tracker = ConnectivityApi24.ConnectivityTrackerCallback { _, _ ->
+            count++
+        }
+        assertEquals(0, count)
+
+        tracker.onAvailable(info)
+        assertEquals(0, count)
+
+        tracker.onAvailable(info)
+        assertEquals(1, count)
+
+        tracker.onAvailable(info)
+        assertEquals(2, count)
+    }
+
+    @Test
+    fun connectivityApi24OnUnvailable() {
+        var count = 0
+        val tracker = ConnectivityApi24.ConnectivityTrackerCallback { _, _ ->
+            count++
+        }
+        assertEquals(0, count)
+
+        tracker.onUnavailable()
+        assertEquals(0, count)
+
+        tracker.onUnavailable()
+        assertEquals(1, count)
+
+        tracker.onUnavailable()
+        assertEquals(2, count)
     }
 }